refactor: move SSOAuthProvider to providers/aws directory
